pg_interface: refactor commit_reveal pg function
